Add basis for custom overlays
The overlay is shown when you click on an editable item and should
contain fields that are editable. This change adds some quick and
dirty code that allows you to customise what is shown on the overlay.

Currently, you can upload and image and change the image and rotation
in the detail section.