Add functionality to inject jquery/cheerio
We've changed the templating JS so that it uses OOP, which means
that we can now inject jquery/cheerio into our templater and
template model. This alleviates the need to include jquery/cheerio
in every single method call that needs it.

We still need to draw a line between injecting dependencies in the
constructor vs declaring dependencies via requirejs. We do know that
it is difficult to include cheerio preloaded with HTML as a
dependency with requireJS, or at least we haven't yet figured out a
way yet, as shim isn't fully supported in node.js.