Clear requirejs cache when generating HTML
Previously, we used the same requirejs object when generating HTML. requirejs
also caches the scripts it loads, so if there are any changes to the frontend
JS scripts and you triggered a regeneration, it would regenerate old content.

Now, we clear the nodejs cache entry for requirejs so that we use a new
requirejs object on each request. This is easier than trying to clear the
requirejs' cache.